+++ bot.py
@@ -12,7 +12,6 @@ load_dotenv()
 TOKEN: Final[str] = os.getenv('DISCORD_TOKEN')
 SERVER_ID = os.getenv('SERVER_ID')
 CHANNEL_ID = os.getenv('CHANNEL_ID')
-print(TOKEN)
 
 #! BOT SETUP
 """
@@ -21,7 +20,7 @@ print(TOKEN)
         to be able to see the message and respond
     
     """
-intents: Intents = Intents.all()
+intents: Intents = Intents.default()
 intents.message_content = True # NOQA
 client: Client = Client(intents=intents) # intents from client
 
@@ -48,36 +47,39 @@ async def send_message(message: Message, user_message: str) -> None:
     except Exception  as e:
         print(e)
 
-#! WEEKLY MESSAGE
 def seconds_until_friday_10pm():
     now = datetime.now()
     next_friday_10pm = (now + timedelta((4 - now.weekday() + 7) % 7)).replace(hour=22, minute=0, second=0, microsecond=0)
-    return (next_friday_10pm - now).total_seconds()
+    seconds = (next_friday_10pm - now).total_seconds()
+    return seconds if seconds >= 0 else seconds + 604800
 
 def seconds_until_saturday_11am():
     now = datetime.now()
     next_saturday_11am = (now + timedelta((5 - now.weekday() + 7) % 7)).replace(hour=11, minute=0, second=0, microsecond=0)
-    return (next_saturday_11am - now).total_seconds()
+    seconds = (next_saturday_11am - now).total_seconds()
+    return seconds if seconds >= 0 else seconds + 604800
 
-@tasks.loop(seconds=seconds_until_friday_10pm())
-async def friday_10pm_task():
-    channel = client.get_channel(CHANNEL_ID)
-    await channel.send("12PM - ISA - PROJECT TIME - ATTANDANCE MANDATORY")
-    friday_10pm_task.change_interval(seconds=seconds_until_friday_10pm())  # Reschedule for next Friday
+#TODO: Fix this bug on channel not found
+# @tasks.loop(seconds=seconds_until_friday_10pm())
+# async def friday_10pm_task():
+#     channel = client.get_channel(CHANNEL_ID)
+#     await channel.send("12PM - ISA - PROJECT TIME - ATTANDANCE MANDATORY")
+#     friday_10pm_task.change_interval(seconds=seconds_until_friday_10pm())  # Reschedule for next Friday
 
-@tasks.loop(seconds=seconds_until_saturday_11am())
-async def saturday_11am_task():
-    channel = client.get_channel(CHANNEL_ID)
-    await channel.send("REMINDER - 12PM - ISA - PROJECT TIME - ATTANDANCE MANDATORY")
-    saturday_11am_task.change_interval(seconds=seconds_until_saturday_11am())  # Reschedule for next Saturday
+#TODO: Fix this bug on channel not found
+# @tasks.loop(seconds=seconds_until_saturday_11am())
+# async def saturday_11am_task():
+#     channel = client.get_channel(CHANNEL_ID)
+#     await channel.send("REMINDER - 12PM - ISA - PROJECT TIME - ATTANDANCE MANDATORY")
+#     saturday_11am_task.change_interval(seconds=seconds_until_saturday_11am())  # Reschedule for next Saturday
   
     
 #! START BOT MESSAGE
 @client.event
 async def on_ready():
     print(f'{client.user} has connected to Discord!')
-    friday_10pm_task.start()
-    saturday_11am_task.start()
+    # friday_10pm_task.start()
+    # saturday_11am_task.start()
     guild = client.get_guild(SERVER_ID)  
     if guild is not None:
         channel = guild.get_channel(CHANNEL_ID) 
@@ -92,11 +94,11 @@ async def on_message(message: Message) -> None:
     if message.author == client.user: # this means bot is the who wrote the message
         return
 
-    usernam: str = str(message.author)
+    username: str = str(message.author)
     user_message: str = message.content
     channel: str = str(message.channel)
     
-    print(f'{usernam} said {user_message} in {channel}')
+    print(f'{username} said {user_message} in {channel}')
     print(f'Processing message: {user_message}')  
     await send_message(message, user_message) # send the message to the send_message function
